feat(glog): 日志文件时间粒度单元测试
